newlib: move `generic` module definitions
Move the `generic` module into the top-level `newlib` module. They were
already reexported by all supported targets in their respective modules.
These types were also conditionally defined anyway.
