+++ src/unix/newlib/horizon/mod.rs
@@ -264,8 +264,3 @@ extern "C" {
 
     pub fn gethostid() -> c_long;
 }
-
-pub use crate::unix::newlib::generic::{
-    dirent,
-    stat,
-};
