+++ src/unix/newlib/mod.rs
@@ -323,6 +323,36 @@ s! {
         // Unverified
         size: [u8; crate::__SIZEOF_PTHREAD_CONDATTR_T],
     }
+
+    #[cfg(all(not(target_os = "vita"), not(target_os = "horizon")))]
+    pub struct sigset_t {
+        __val: u32,
+    }
+
+    #[cfg(not(target_os = "vita"))]
+    pub struct stat {
+        pub st_dev: crate::dev_t,
+        pub st_ino: crate::ino_t,
+        pub st_mode: crate::mode_t,
+        pub st_nlink: crate::nlink_t,
+        pub st_uid: crate::uid_t,
+        pub st_gid: crate::gid_t,
+        pub st_rdev: crate::dev_t,
+        pub st_size: crate::off_t,
+        pub st_atim: crate::timespec,
+        pub st_mtim: crate::timespec,
+        pub st_ctim: crate::timespec,
+        pub st_blksize: crate::blksize_t,
+        pub st_blocks: crate::blkcnt_t,
+        pub st_spare4: [c_long; 2usize],
+    }
+
+    #[cfg(not(target_os = "vita"))]
+    pub struct dirent {
+        pub d_ino: crate::ino_t,
+        pub d_type: c_uchar,
+        pub d_name: [c_char; 256usize],
+    }
 }
 
 pub const PTHREAD_MUTEX_INITIALIZER: pthread_mutex_t = pthread_mutex_t {
@@ -933,8 +963,6 @@ extern "C" {
     pub fn popen(command: *const c_char, mode: *const c_char) -> *mut crate::FILE;
     pub fn uname(buf: *mut crate::utsname) -> c_int;
 }
-
-mod generic;
 
 cfg_if! {
     if #[cfg(target_os = "espidf")] {
