+++ src/unix/newlib/powerpc/mod.rs
@@ -3,12 +3,6 @@ use crate::prelude::*;
 pub type clock_t = c_ulong;
 pub type wchar_t = c_int;
 
-pub use crate::unix::newlib::generic::{
-    dirent,
-    sigset_t,
-    stat,
-};
-
 // the newlib shipped with devkitPPC does not support the following components:
 // - sockaddr
 // - AF_INET6
